pyproject.toml
[tool.poetry]
name = "ascii-art"
version = "0.1.0"
description = "Presents images using ASCII characters."
authors = ["Zero to Mastery <admin@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^8.3.2"
pyfiglet = "^0.8.post1"
Flask = "^2.0.2"
colorama = "^0.4.4"
termcolor = "^1.1.0"
python-dotenv = "^0.19.1"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
